port flatten
